third_party/WebKit/LayoutTests/inspector/timeline/timeline-script-tag-1-expected.txt

CONSOLE MESSAGE: line 8: iframe loaded
Tests the Timeline API instrumentation of an HTML script tag.


ParseHTML
ParseHTML
----> EvaluateScript
--------> Timestamp: SCRIPT TAG
EvaluateScript Properties:
{
    children : <object>
    data : {
        lineNumber : <number>
        url : .../inspector/timeline/resources/timeline-iframe-data.html
    }
    endTime : <number>
    frameId : <string>
    startTime : <number>
    thread : <string>
    type : "EvaluateScript"
}
ParseHTML
ParseHTML
----> EventDispatch
--------> FunctionCall
------------> FunctionCall
----> MarkLoad
----> MarkDOMContent
